Refactor Snapshots
